refactor(extensions): rename content category to transform
- src/extensions/content/ → src/extensions/transform/
- extensions/ext-content-mdx/ → extensions/ext-transform-mdx/
- Package name @veryfront/ext-content-mdx → @veryfront/ext-transform-mdx
- Runtime extension name "ext-content-mdx" → "ext-transform-mdx"
- deno.json workspace + ./extensions/transform export + imports alias
- Consumer sweep: builtin-extensions.ts, cli/shared/, transforms/mdx/,
  recommendations.ts, ext-transform-mdx internals (runtime name,
  JSDoc, log labels, tests)

Rationale: aligns the category folder name with the underlying
ContentTransformer interface and keeps the ext-{category}-{subject}
prefix pattern consistent across the workspace.
