+++ deno.json
@@ -10,11 +10,11 @@
     "./extensions/ext-auth-jwt",
     "./extensions/ext-bundler-esbuild",
     "./extensions/ext-cache-redis",
-    "./extensions/ext-content-mdx",
     "./extensions/ext-css-tailwind",
     "./extensions/ext-node-compatibility",
     "./extensions/ext-parser-babel",
-    "./extensions/ext-tracing-opentelemetry"
+    "./extensions/ext-tracing-opentelemetry",
+    "./extensions/ext-transform-mdx"
   ],
   "exclude": [
     "npm/",
@@ -73,12 +73,12 @@
     "./extensions/bundler": "./src/extensions/bundler/index.ts",
     "./extensions/cache": "./src/extensions/cache/index.ts",
     "./extensions/compat": "./src/extensions/compat/index.ts",
-    "./extensions/content": "./src/extensions/content/index.ts",
     "./extensions/css": "./src/extensions/css/index.ts",
     "./extensions/database": "./src/extensions/database/index.ts",
     "./extensions/parser": "./src/extensions/parser/index.ts",
     "./extensions/schema": "./src/extensions/schema/index.ts",
     "./extensions/tracing": "./src/extensions/tracing/index.ts",
+    "./extensions/transform": "./src/extensions/transform/index.ts",
     "./testing": "./src/testing/index.ts",
     "./testing/assert": "./src/testing/assert.ts",
     "./testing/bdd": "./src/testing/bdd.ts",
@@ -163,13 +163,13 @@
     "veryfront/testing/bdd": "./src/testing/bdd.ts",
     "veryfront/extensions/bundler": "./src/extensions/bundler/index.ts",
     "veryfront/extensions/compat": "./src/extensions/compat/index.ts",
-    "veryfront/extensions/content": "./src/extensions/content/index.ts",
     "veryfront/extensions/contracts": "./src/extensions/contracts.ts",
     "veryfront/extensions/css": "./src/extensions/css/index.ts",
     "veryfront/extensions/database": "./src/extensions/database/index.ts",
     "veryfront/extensions/parser": "./src/extensions/parser/index.ts",
     "veryfront/extensions/schema": "./src/extensions/schema/index.ts",
     "veryfront/extensions/tracing": "./src/extensions/tracing/index.ts",
+    "veryfront/extensions/transform": "./src/extensions/transform/index.ts",
     "jose": "npm:jose@5.9.6",
     "#veryfront": "./src/index.ts",
     "#veryfront/agent": "./src/agent/index.ts",
