+++ extensions/ext-transform-mdx/src/plugins/rehype-mermaid.ts
